Added a few extra checks

+++ pysims/job_controller.py
@@ -219,9 +219,10 @@ def submit_jobs(
     }
 
     case_dir = os.path.join(WORKING_DIR, f"cases/{job_name}_case")
-    if os.path.exists(case_dir):
-        shutil.rmtree(case_dir)
-    os.mkdir(case_dir)
+    if checkpoint_file is None:
+        if os.path.exists(case_dir):
+            shutil.rmtree(case_dir)
+        os.mkdir(case_dir)
 
     context = JobContext(job_name, config_generator, case_dir, ext, andromeda2_get_partition(time), nodes, ncores, memory, time, cleanup, metaparams)
